pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robotmcp-server"
version = "2.0.0"
description = "Modular MCP server with Streamable HTTP, optional OAuth 2.1, Supabase auth, Cloudflare tunnel, and creator-only access control"
readme = "README.md"
license = {text = "Proprietary - Copyright (c) 2025 Contoro. All rights reserved."}
requires-python = ">=3.10"
keywords = ["mcp", "oauth", "supabase", "cloudflare", "chatgpt", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastmcp>=2.14.0",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.26,<0.29",
    "requests>=2.31.0",
    "PyJWT>=2.10.1",
    "cryptography>=44.0.0",
    "python-multipart>=0.0.17",
    "supabase>=2.25.0",
]

[project.scripts]
robotmcp-server = "cli:main"

[project.urls]
Homepage = "https://github.com/robotmcp/robotmcp_server"
Repository = "https://github.com/robotmcp/robotmcp_server"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "oauth/*.py",
    ".env.public",
]

[tool.hatch.build.targets.sdist]
include = [
    "*.py",
    "oauth/*.py",
    ".env.public",
    "requirements.txt",
    "README.md",
]
